Update application_stop.sh
Updated kill command

+++ scripts/application_stop.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
 #Stopping existing node servers
 echo "Stopping any existing Node servers"
-pkill node
+if pgrep node; then pkill node; fi
